requirements.txt
cryptography==41.0.7
PyPDF2==3.0.1
python-dotenv==1.0.0
tabulate==0.9.0
